+++ NEWS.md
@@ -1,6 +1,7 @@
 # explore 1.2.1
 
 * add parameter `nthread` to `explain_xgboost()` (#45)
+* fix undefined/not meaningful values in abtest() shiny-app
 
 # explore 1.2.0